+++ packages/relay/src/lib/types/events.ts
@@ -1,6 +1,29 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import { IEthExecutionEventPayload, IExecuteQueryEventPayload, IExecuteTransactionEventPayload } from './sdkClient';
+import { RequestDetails } from './RequestDetails';
+
+export interface IEthExecutionEventPayload {
+  method: string;
+}
+
+export interface IExecuteQueryEventPayload {
+  executionMode: string;
+  transactionId: string;
+  txConstructorName: string;
+  cost: number;
+  gasUsed: number;
+  status: string;
+  requestDetails: RequestDetails;
+  originalCallerAddress: string | undefined;
+}
+
+export interface IExecuteTransactionEventPayload {
+  transactionId: string;
+  txConstructorName: string;
+  operatorAccountId: string;
+  requestDetails: RequestDetails;
+  originalCallerAddress: string;
+}
 
 export interface TypedEvents {
   eth_execution: [IEthExecutionEventPayload];
